requirements-test.txt
biopython==1.64
coverage==3.7.1
coveralls==0.4.4
docutils==0.12
nose==1.3.4
Sphinx==1.2.3
